Simplify bytecode iterator, all cases are the same now
